Added calculator with brain and display and operations enum

+++ src/org/academiadecodigo/calculator/Display.java
@@ -5,7 +5,7 @@ package org.academiadecodigo.calculator;
  */
 public class Display {
 
-    public void showResult(int result) {
+    public void showResult(double result) {
         System.out.println(result);
     }
